requirements.txt
gymnasium
torch
pandas
matplotlib
shapely
tensordict
torchrl
pydantic_settings
python-dotenv
tqdm